+++ ContentValidation/Validation.cs
@@ -1,0 +1,15 @@
+namespace UtilityLibraries;
+public interface IValidation{
+    Task<TResult> Validate(string testLink);
+}
+
+public class TResult
+{
+    public bool Result { get; set; }
+    public string? ErrorMsg { get; set; }
+
+    public TResult(){
+        Result = true;
+        ErrorMsg = "";
+    }
+}
